# docs/swagger.yaml
definitions:
...
  entity.TransactionResponse:
    properties:
      category:
        description: DateTransaction time.Time `jsons:"dateTransaction"`
        type: string
      createdAt:
        type: string
      id:
        type: integer
      nameTransaction:
        type: string
      updatedAt:
        type: string
      value:
        type: integer
    type: object
  entity.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      profission:
        type: string
      updatedAt:
        type: string
    type: object
...
  handler.CreateTransactionRequest:
    properties:
      category:
        description: DateTransaction time.Time `jsons:"dateTransaction"`
        type: string
      nameTransaction:
        type: string
      value:
        type: integer
    type: object
  handler.CreateTransactionResponse:
    properties:
      data:
        $ref: '#/definitions/entity.TransactionResponse'
      message:
        type: string
    type: object
  handler.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      profission:
        type: string
    type: object
  handler.CreateUserResponse:
    properties:
      data:
        $ref: '#/definitions/entity.UserResponse'
      message:
        type: string
    type: object
...
  handler.DeleteTransactionResponse:
    properties:
      data:
        $ref: '#/definitions/entity.TransactionResponse'
      message:
        type: string
    type: object
  handler.DeleteUserResponse:
    properties:
      data:
        $ref: '#/definitions/entity.UserResponse'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
...
  handler.GetAllTransactionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.TransactionResponse'
        type: array
      message:
        type: string
    type: object
  handler.GetAllUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/entity.UserResponse'
        type: array
      message:
        type: string
    type: object
...
  handler.ShowTransactionResponse:
    properties:
      data:
        $ref: '#/definitions/entity.TransactionResponse'
      message:
        type: string
    type: object
  handler.ShowUserResponse:
    properties:
      data:
        $ref: '#/definitions/entity.UserResponse'
      message:
        type: string
    type: object
...
  handler.UpdateTransactionRequest:
    properties:
      category:
        type: string
      nameTransaction:
        type: string
      value:
        type: integer
    type: object
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      profission:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /transaction:
    delete:
      consumes:
      - application/json
      description: Delete a new Transaction
      parameters:
      - description: Transaction identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Transactions
    get:
      consumes:
      - application/json
      description: Show a new Transaction
      parameters:
      - description: Transaction identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Transactions
    post:
      consumes:
      - application/json
      description: Create a new Transaction
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Transactions
    put:
      consumes:
      - application/json
      description: Update an Transaction
      parameters:
      - description: Transaction identification
        in: query
        name: id
        required: true
        type: string
      - description: Transaction data to Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Transactions
  /transactions:
    get:
      consumes:
      - application/json
      description: List all Transactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetAllTransactionResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Transactions
  /user:
    delete:
      consumes:
      - application/json
      description: Delete a new User
      parameters:
      - description: User identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Show a new User
      parameters:
      - description: User identification
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ShowUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new User
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update an User
      parameters:
      - description: User identification
        in: query
        name: id
        required: true
        type: string
      - description: User data to Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Users
  /users:
    get:
      consumes:
      - application/json
      description: List all User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GetAllUserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - Users
swagger: "2.0"